requirements.txt
streamlit>=1.32
streamlit-folium>=0.18
folium>=0.16
plotly>=5.20
pandas>=2.2
pyarrow>=15
httpx>=0.27
pyyaml>=6.0
geopandas>=0.14
pytest>=8.0
numpy>=1.26
